Refactor telegram handlers. Fix linting
